+++ src/snarky.d.ts
@@ -696,6 +696,7 @@ declare const Test: {
     serializeCommon(common: string): { data: Uint8Array };
     hashPayment(payment: string): string;
     hashPaymentV1(payment: string): string;
+    hashZkAppCommand(command: string): string;
   };
 };